requirements.txt
yt-dlp>=2025.1.15
pytesseract
pillow
google-generativeai